+++ app/Http/Resources/CommentResource.php
@@ -21,7 +21,9 @@ class CommentResource extends JsonResource
             'created_at' => $this->created_at->format('Y-m-d H:i:s'),
             'updated_at' => $this->updated_at->format('Y-m-d H:i:s'),
             'num_of_reactions' => $this->reactions_count,
+            'num_of_comments' => $this->comments_count,
             'current_user_has_reaction' => $this->reactions->count() > 0,
+            'comments' => CommentResource::collection($this->comments),
             'user' => [
                 'id' => $this->user->id,
                 'name' => $this->user->name,
