+++ app/Models/Comment.php
@@ -5,13 +5,14 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 class Comment extends Model
 {
     use HasFactory;
 
-    protected $fillable = ['post_id', 'comment', 'user_id'];
+    protected $fillable = ['post_id', 'comment', 'user_id', 'parent_id'];
 
     public function user(): BelongsTo
     {
@@ -27,4 +28,8 @@ class Comment extends Model
         return $this->morphMany(Reaction::class,'object');
     }
 
+    public function comments(): HasMany
+    {
+        return $this->hasMany(self::class, 'parent_id');
+    }
 }
